+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changed the default scorer for relevancy searches to DISMAX and added a filter (`redipress/scorer`) for changing that.
 - `post_mime_type` field to the core schema and the functions to support querying it.
 
+### Fixed
+- Singular view main queries where there is `paged` parameter present at the same time.
+
 ## [1.4.1] - 2020-05-20
 
 ### Fixed