+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2020-06-05
+
+### Added
+- A reverse filter for getting the Search class instance.
+
+### Fixed
+- A bug where `"posts_per_page" => -1` query argument would result in empty result set.
+
 ## [1.5.1] - 2020-06-04
 
 ### Added